lib: new state function netplan_state_update_yaml_hierarchy
This new function basically refreshes all the configuration files that
have been used to create the current state. This is where the global
source file tracking comes into play, as it allows us to delete config
files that didn't result in any actual settings in the final state, i.e.
because their defs were removed in a subsequent patch.

This function will be used in the netplan set rewrite.

V2:
* Do some default_filename early sanity checks
* Error out if removal of obsolete files fails.