parser: partial support of null fields
This support is done by pre-processing the input files to identify the
null fields, and when we find a key that matches a null field we simply
skip to the next key.

This necessitates to keep track of the complete key throughout the
parser, hence the relatively big size of the patch.

V2:
* Rename prefix arguments to key_prefix
* Use \t rather than . as prefix segment separator