libnetplan: bump the XOPEN_SOURCE definition to 700
This allows us to access safer string manipulation functions such as
stpncpy (no that's not a type), which are available in the POSIX 2008
standard. This specific function will be useful in implementing the new
API pattern of user-provided string buffers.

In the newer versions of the POSIX standard, usleep has been deprecated
then removed entirely, to be replaced by nanosleep, hence the change in
dbus.c.