!API! lib: netdef: rename filename into filepath
As discussed, we rename the ambiguous filename field into filepath, as
it denotes a full file patch, not the file base name. We introduce a new
API for it that conforms to the API design guidelines (that should
really be spelled out in the repo :-) ).

The older function is kept around for ABI compatibility but moved out of the
API, hence the API break warning.