manpages: fix long options rendered with an en-dash instead of two dash
In generated manpages, the long options look as if they were prefixed
with a single-dash "-", as in the following example:

 netplan [–debug] generate [–root-dir ROOT_DIR] [–mapping MAPPING]

In reality, the double-dash notation was replaced by an en-dash
(U+2013) when generating the manpage.

This is caused by the "smart" extension that pandoc enables by default
when reading markdown.

Fixed by preventing pandoc from enabling the smart extension when
generating manpages. This is only possible when we explicitly specify
the input format.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>