tests: configmanager: Fix inconsistencies in test data
There are a couple of inconsistencies in the test data. The test.yaml
netplan config tries to configure a vlan interface on eth99, which
hasn't been defined, and the merged config tries to add two new links on
patcha and patchb, while they're already in a link together. While the
current code doesn't catch this, parser consolidation will result in
stricter validation, making the merge test fail.