lib: move complex_object_is_dirty to utils with better API
Instead of forcing users to cast (which was done through a macro in
netplan.c anyway), we now use the proper type to say "give us a pointer
to whatever, we promise we won't touch its data".

This function will be used to expose a bit of the dirtiness of the bonds
and bridges properties to Python, which in turn will tell us how
revertable a given config is.