test:cli:set: add regression cases for LP: #1997467
When parsing the full YAML hierarchy (from 'netplan set'), new settings might
not end up in the "to_write" GList (and therefore not in the new origin-hint
file), as the same interface might have been defined in another, previous file
(e.g. 0-snapd-defaults.yaml) and therefore any updates are redirected to that
previous file, or ignored if a origin-hint is passed, which does not match
the previous filename (e.g. origin-hint: 90-snapd-config.yaml).

Still we need to parse the full YAML hierarchy for validation reasons, as a new
origin-hint file might describe new definitions of an interface that contains
references to a netdef from another YAML file. E.g.:

$ cat 0-snapd-defaults.yaml
network:
  ethernets:
    eth0:
      dhcp4: true

$ netplan set --origin-hint=90-snapd-config "bridges.br54.interfaces=[eth0]"

=> This would fail if we wouldn't parse the full YAML hierarchy, before writing
the 90-snapd-config.yaml file, as it would not be aware of the "eth0" netdef.