Fix a bug in the generate --mapping command
When the parameter --mapping is used with generate, it should only try
to show the interface mapping to the user and exit. As it is now, it
will call the _cleanup() functions first.

Also, if the parameter is used but there is no netdefs in the
configuration, the --mapping code path will be skipped without jumping
to the cleanup label and will return 0.

It fixes LP: #1996941