Migrate from (deprecated) nose to pytest
Pytest and pycoverage require some extra configuration to track test
coverage in tests that spawn subprocesses. The new files .coveragerc and
sitecustomize.py are intended to enable support for subprocesses.

The test set test_terminal.py is completely skipped when testing on my
system and, because of that, coverage test was reporting a low coverage.
I still need to see it running in our automation to validate if we still
need to skip them. Another workaround would be add a nocover to the
Terminal class.

Also:
  - Fix some deprecation warnings in some tests.
  - Remove this task from the TODO list :)