generate:util: fix double-slash root filepath
Make sure we're using proper paths values (i.e. NOT "//etc/netplan/...")
as we're doing string comparisons on those values and the leading
double-slash might confuse those checks.

E.g. in netplan.c:netplan_state_write_yaml_file() we compare netdef->filepath
to the filepath of a new YAML file to be written. If a netdef has been parsed
but its filepath is using double slashes, it won't match the new file and the
new file will be deleted instead of updated.

Starting with an empty Netplan config in /etc/netplan:
$ netplan set --origin-hint test "bridges.br54.dhcp4=false"
=> will create test.yaml, containing the br54 definition
$ netplan set --origin-hint test "bridges.br54.dhcp4=true"
=> updating the same br54 netdef (re-using 'netplan set') will delete test.yaml