src:parse:netplan: write global renderer depending on output file
When parsing the global renderer, we also want to store the path+filename of
the parsed value in a new ->global_renderer datastructure (in parser & state).
This way, when generating YAML output, we can put a different/new renderer
stanza (e.g. from "netplan set network.renderer=...") into a higher order YAML
file while conserving the renderer data in the original YAML files.

The renderer can be parsed from an existing file or a new, unnamed YAML patch.
If the path is known, we want to write back the data to that file only. If the
path is unknown/anonymous (i.e. emtpy string) we want to write that data to the
fallback file (e.g. 70-netplan-set.yaml or <origin-hint>.yaml).