util: Add NetplanStateIterator type and functions
The new NetplanStateIterator struct keeps the iterator state
(basically the front of the netdefs_ordered list) and
netplan_state_iterator_next() will return the next netdef in the list
and update the iterator to point to the next netdef.

Sometimes it is useful to iterate through netdefs, specially when you
don't know the netdef ID. One first consumer for this API would probably
be the NetworkManager + Netplan integration. When NetworkManager loads
the keyfile to the netplan state, we might don't know if the netdef ID
is the interface name or the NM-<UUID> string. With this API we don't
need to know, we can just get the first netdef in the list. As the new
Netplan API doesn't expose internal structures, some API like this one
is necessary.