generator: Don't reload udev in the generator
The systemd documentation states that any kind of process communication
shouldn't happen inside generators. It can cause problems like the one
reported in LP#1999178.

Ubuntu seems to not be affected by the same issue yet, but as the
problem happens due to a race condition we might run into it in the
future as well.

This change moves the call to the udevadm command from the generator to
the Python code. To keep the behaviour as close to the previous one as
possible, the call is made as part of the generate command.

We could simply check if the generator is being called as a systemd
generator and don't call udevadm, but in the end the binary still is a
generator so we shouldn't even have that logic there.

The same is true for the calls we make to systemd using systemctl. Even
though they are not executed when the binary is called as a generator,
they probably shouldn't be there.