Netplan api iterator (#306)
The Netplan state iterator API is intended to enable consumers to walk through the ordered list of network definitions. It is useful, for instance, when you need to save the netdefs to files but you don't know the netdef ID to lookup for it.

The idea of having a non-opaque struct exposed to consumers was borrowed from Glib ([here](https://github.com/GNOME/glib/blob/abd76e02864163b6a627e5c5722cb4d6a8d35501/glib/ghash.h#L46) and [here](https://github.com/GNOME/glib/blob/abd76e02864163b6a627e5c5722cb4d6a8d35501/glib/ghash.c#L1143)). By doing this, the iterator can be placed at the stack, which is much more convenient.

This API was integrated and tested with the NetworkManager+Netplan integration code base.

COMMITS:
* util: Add NetplanStateIterator type and functions

The new NetplanStateIterator struct keeps the iterator state
(basically the front of the netdefs_ordered list) and
netplan_state_iterator_next() will return the next netdef in the list
and update the iterator to point to the next netdef.

Sometimes it is useful to iterate through netdefs, specially when you
don't know the netdef ID. One first consumer for this API would probably
be the NetworkManager + Netplan integration. When NetworkManager loads
the keyfile to the netplan state, we might don't know if the netdef ID
is the interface name or the NM-<UUID> string. With this API we don't
need to know, we can just get the first netdef in the list. As the new
Netplan API doesn't expose internal structures, some API like this one
is necessary.

* state iterator: Don't use an opaque struct

In order to enable consumers to place the interator at the stack, the
type can't be an opaque struct. The idea is based on glib's
implementation of iterators.

Also, use some glib helpers for simplicity and clarity when accessing
GList elements.