networkd: make sure VXLAN is in the right section
If an interface has a list of routes and is linked to a vxlan, the
VXLAN keyword is added to the [Route] section in the resulting networkd
configuration file.

This change makes the vxlan handling to happen before routes so VXLAN=
will be added to the [Network] section.

It fixes LP#2000713