vxlan: initialize mac_learning with 1
Vxlan learning is enabled by default on Linux. On Netplan it's a boolean
variable that we initialize with false and only include in the generated
configuration if it's true. That means we can't disable it on vxlan
interfaces even if we define mac-learning: false explicitly in the yaml
file.

That is not the ideal solution! The proper way to do that is defining it
as a tristate variable but it requires changes in the netplan_vxlan
structure. One bad side effect of this change is that the yaml
configuration generated from netdefs and the NM/networkd configuration
will contain the option mac-learning even if we don't have it in the
original yaml.

This addresses LP#2000712