cli:set: fix origin-hint handling (LP: #1997467)
When using the new libnetplan YAML parser, we need to ignore any global values
(like "renderer") or netdefs (that are part of a given YAML patch) when parsing
the existing YAML hierarchy, unless they're about to be parsed from an existing
origin-hint file of the same name.

That is to avoid assigning updates to those stanzas/netdefs to existing network
definitions inside other files (not the origin-hint), as this would lead to
those updates being ignored when we're writing out just the single origin-hint
file (using state.write_yaml_file()). We still need to parse and validate the
full context of other netdefs (not part of a given YAML patch), as the patch
might reference existing netdefs (e.g. as bridge interfaces), or change current
configuration in a way that makes it invalid (e.g. changing vrf tables),
therefore they need to exist in the (validation-)parser's context. Also we
need to parse such netdefs from an origin-hint file of same name, as we'd want
update the definition in that file (instead of overriding it).

We do this by adding relevant global values and netdefs, which are part of the
YAML patch, to the "null_overrides" data, before parsing the full YAML
hierarchy. Adding the filename of the origin-hint as a constraint, which can be
checked when parsing to not ignore relevant netdefs from an exsisting
origin-hint file.