parser: warning if the interface name is too long
Interfaces names are limited to 15 characters (see IF_NAMESIZE in
net/if.h). Currently Netplan will allow the user to use invalid names
without a warning. In this case the operation will be ignored by the
backends.

With this change, Netplan will show a warning but still complete
successfully. As there might be users with invalid interface names in
their configuration without knowing it, changing the behavior to fail
would probably cause problems.