doc/examples: remove unnecessary route for IPv6 on-link gateways
When the on-link flag is enabled, the additional route has no effect.
This extra route is only required when the on-link flag is *not* set.

Indeed, for netorkd, when "on-link" is true, the resulting
systemd-networkd config file contains "GatewayOnLink=true", which
translates to "flags: onlink" in its logs. The resulting route appears
with this "onlink" flag in the output of "ip route".

Similarly, for NetworkManager, "route1_options=onlink=true" gets added
by
https://github.com/canonical/netplan/blob/e5ff9f6dbeb20a8a3e6abc712efcf1b0008afa75/src/nm.c#L228-L230