Check if the interface name is too long (LP: #1988749) (#313)
* tests: fix a keyfile test logic for bridges

For bridges, the emitter will use the interface name as ID and not the
NetworkManager UUID.

* parser: warning if the interface name is too long

Interfaces names are limited to 15 characters (see IF_NAMESIZE in
net/if.h). Currently Netplan will allow the user to use invalid names
without a warning. In this case the operation will be ignored by the
backends.

With this change, Netplan will show a warning but still complete
successfully. As there might be users with invalid interface names in
their configuration without knowing it, changing the behavior to fail
would probably cause problems.

* validation: check if netdef is NULL

Also, fix apply.py code style to make the linter happy.

* ctests: add tests for the new validation function

Also add a utility to load the yaml content from strings.

* validation: add a comment about IF_NAMSIZE warning

* nm: replace undocumented constant with IF_NAMESIZE

* ctest:utils: adopt to netplan_error_clear rename

Co-authored-by: Lukas Märdian <slyon@ubuntu.com>