vxlan: convert some settings to tristate (LP: #2000712) (#311)
Vxlan learning is enabled by default on Linux. On Netplan it's a boolean variable that we initialize with false and only include in the generated configuration if it's true. That means we can't disable it on vxlan interfaces even if we define mac-learning: false explicitly in the yaml file.

This PR converts the settings mac-learning, arp-proxy and short-circuit from gboolean to tristate.

This addresses LP#2000712

COMMITS:

* vxlan: initialize mac_learning with 1

Vxlan learning is enabled by default on Linux. On Netplan it's a boolean
variable that we initialize with false and only include in the generated
configuration if it's true. That means we can't disable it on vxlan
interfaces even if we define mac-learning: false explicitly in the yaml
file.

That is not the ideal solution! The proper way to do that is defining it
as a tristate variable but it requires changes in the netplan_vxlan
structure. One bad side effect of this change is that the yaml
configuration generated from netdefs and the NM/networkd configuration
will contain the option mac-learning even if we don't have it in the
original yaml.

This addresses LP#2000712

* Revert "vxlan: initialize mac_learning with 1"

This reverts commit 94e9c227bb9728ff1ad1cda3ebe407ab80ae5181.

* parser: convert some vxlan options to tristate

Settings mac-learning, arp-proxy and short-circuit were converted from
gboolean to tristate.

When these options are enabled by default in the backend, it's not
possible to disable them via Netplan as it was including the option in
the generated configuration only when it was set to true.

It addresses LP#2000712

* parser: remove unused function

* tests: reduce duplication with config templates