Match by PermanentMACAddress (#278)
Using bond, bridge, and VLAN devices results in sharing MAC addresses across interfaces. Matching by MACAddress fails in two ways. It will overmatch by matching all the interfaces (physical and virtual) that share that MAC address. It will undermatch by no longer matching the physical interface once its MAC address has been changed.

Matching by PermanentMACAddress fixes both of those issues.

This replaces the old "workaround" of setting Type=!vlan bond bridge.

That code had a comment saying, "something outputs netplan config that includes using the MAC of the first phy member of a bond as default value for the MAC of the bond device itself." I think this was a misunderstanding of what was going on there. In both examples, the configuration was attempting to match a physical interface by MAC address, not provide a MAC address to set on the virtual device.

These are the two Launchpad bugs referenced by the code:
https://bugs.launchpad.net/netplan/+bug/1804861
https://bugs.launchpad.net/ubuntu/+source/systemd/+bug/1888726

The comment continued, "This is evil, it's an optional field and networkd knows what to do if the MAC isn't specified". This configuration is perfectly reasonable. It is in fact required, as explained in the netplan documentation, if one wants to use settings that apply in the .link file (wakeonlan or the offload settings). This is because networkd refuses to apply a .link file that matches by name if the interface has already been renamed. At a minimum, this means that matching by MAC address is required if one wants to be able to change the settings after boot (on interfaces that are renamed). If the interfaces are renamed at early boot, before the netplan configuration applies, matching by MAC address is required for those settings to apply at all.

For my example, see:
#270 (comment)
#270 (comment)

Arguably, netplan should always use PermanentMACAddress, at least for physical interfaces. But that change might have unintended consequences, so I opted for the more minimal change of only using it in these complex scenarios (which are already special-cased).

COMMITS:

* Match by PermanentMACAddress in complex scenarios

Using bond, bridge, and VLAN devices results in sharing MAC addresses
across interfaces.  Matching by MACAddress fails in two ways.  It will
overmatch by matching all the interfaces (physical and virtual) that
share that MAC address.  It will undermatch by no longer matching the
physical interface once its MAC address has been changed.

Matching by PermanentMACAddress fixes both of those issues.

This replaces the old "workaround" of setting "Type=!vlan bond bridge".

That code had a comment saying, "something outputs netplan config that
includes using the MAC of the first phy member of a bond as default
value for the MAC of the bond device itself."  I think this was a
misunderstanding of what was going on there.  In both examples, the
configuration was attempting to match a physical interface by MAC
address, not provide a MAC address to set on the virtual device.

The comment continued, "This is evil, it's an optional field and
networkd knows what to do if the MAC isn't specified".  This
configuration is perfectly reasonable.  It is in fact required,
as explained in the netplan documentation, if one wants to use settings
that apply in the .link file (wakeonlan or the offload settings).
This is because networkd refuses to apply a .link file that matches by
name if the interface has already been renamed.  At a minimum, this
means that matching by MAC address is required if one wants to be able
to change the settings after boot (on interfaces that are renamed).  If
the interfaces are renamed at early boot, before the netplan
configuration applies, matching by MAC address is required for those
settings to apply at all.

Arguably, netplan should always use PermanentMACAddress, at least for
physical interfaces.  But that change might have unintended
consequences, so I opted for the more minimal change of only using it
in these complex scenarios (which are already special-cased).

* doc: update match.mac wrt permanent vs virtual MAC

* networkd: apply PermanentMACAddress= to all MAC matches

* test:integration:base: add veth MAC matching quirk

---------

Co-authored-by: Lukas Märdian <slyon@ubuntu.com>