parser: check for route duplicates (LP: #2003061) (#320)
We currently check if the number of routes we have in the list is the same as the number of entries in the yaml structure. This cause problems when routes come from different files, see LP#2003061. In this case, the second route is not added to the netdef.

This patch changes the logic to actually check if the route being processed is already in the list. For the uniqueness check, it compares the route IP family, table ID, metric, from, to and via fields.

COMMITS:

* parser: check for route duplicates

We currently check if the number of routes we have in the list is the
same as the number of entries in the yaml structure. This cause problems
when routes come from different files, see LP#2003061. In this case, the
second route is not added to the netdef.

This patch changes the logic to actually check if the route being
processed is already in the list. For the uniqueness check, it compares
the route table ID, metric, from, to and via fields.

* parser: improve routes comparison

Take into account that the destination could be 'default' or
'0.0.0.0/0' (or '::/0' for IPv6). In this case we should
consider them the same.
Add the family to the matching criteria as well.

* tests: Simplify and add more test cases