ATTN: parser: validate lacp-rate properly
This field should accept only the values "fast" and "slow". It's
documented in the Netplan documentation.

It addresses LP: #1745648

Note for backports/SRUs: this change will make invalid values
that were being accepted before to break netplan generate. While we want
this fixed for future releases, this shouldn't be adopted by stable
releases of Ubuntu. The recommended option for backports is to demote
the yaml_error to a simple warning so the user will know that the
configuration needs to be fixed.