netplan: add "validate" command
Netplan validate can be used to check the configuration before trying to
apply it. It's useful when one is making changes in the configuration
and want to periodically check if it is still valid without having to call
"apply" or "generate".

The validate command will load, parse and validate the configuration
without saving anything to disk.