parser: reimplement optional-addresses
optional-addresses is generating invalid configuration for
systemd-networkd. The option OptionalAddresses doesn't really exist.
It doesn't emit any configuration for Network Manager.

This changes reimplement optional-addresses using systemd's
"RequiredFamilyForOnline" and NetworkManager's "may-fail".

It keeps all the old options, for backwards compatibility, as noops, as
they were before in practice, and introduce 3 new options: "ipv4",
"ipv6" and "none".

On systemd, RequiredFamilyForOnline defaults to "any", meaning that any
interface that gets an IP will be enough to make the system online.
Netplan will emit configuration for systemd only when the value of
optional-addresses is "none", meaning that nothing is optional. In this
case, RequiredFamilyForOnline will have the value "both".

On Network Manager, it will emit "may-fail" as true or false according
to the configuration.

It will still allow the user to have the old values so we will not break
systems out there. A warning message will emitted is this case.

It resolves LP: #1880029