parse: set the backend on nm-devices to NM by default (#349)
* parse: set the backend on nm-devices to NM by default

By definition, the renderer on nm-devices must be NetworkManager. Not
defining it will default to networkd. Apart from the fact it doesn't
make sense, it is causing a memory leak because the function
g_datalist_clear() will no be called in reset_backend_settings() if the
backend is not NM.

* parse.c: formatting