tests: add a new parameter to generate_from_keyfile
The parameter "regenerate" enable the caller to check or not if the
regenerated keyfile matches the input.

This will be used by tests that parse keyfiles with implicit
configuration. In this cases, the regenerated file will include this
configuration and will not match the input. For example, the Wireguard
property allowed-ips accept IPs without a prefix, but the prefix /32 is
implicitly used as it is required by Wireguard.