parse-nm:wg: auto append the prefix to allowed IPs
Network Manager doesn't fail if an IP in the allowed-ips list doesn't
have a prefix. In this case, WG will default to /32.

As the WG's documentation says the prefix must be used, Netplan will
fail the validation step if it's not present. To be able to parse a
keyfile and generate a YAML that will not fail validation, let's append
a /32 automatically to the value if it doesn't have a prefix defined by
the user.