networkd: generate configuration for virtual-ethernets
When generating configuration for veth, only one .netdev file should be
created for the pair. The logic used to select which netdef will be used
to generate the file is simple: we sort the pair names and use the first
one.

If the veth's pair wasn't defined in the YAML file we don't hard fail,
just skip it.