parser: introduce the "placeholder" interface type
This change is intended to be used only by Network Manager, more
specifically when used through the integration with Netplan.

Network Manager doesn't enforce the existence of interfaces in order to
create a connection. For example, when creating a VLAN, it doesn't care
if the parent interface doesn't exist. The nmconnection will be created
anyway.

When creating a VLAN directly in Netplan, the "link" property is
required and the interface used as "link", the parent interface, must
exist in Netplan. It will throw an error otherwise.

Since the integration with Network Manager, some of these restrictions
need to be relaxed or NM will error out when creating these types of
connections.

The NETPLAN_DEF_TYPE_NM_PLACEHOLDER_ is intended to fill the gaps of interfaces
that must exist for the connection to work but don't really exist.

Practical example: when creating a VLAN, the "link" property is required
and the interface pointed by the link also must exist somewhere in the
YAML files. Although, when creating it through Network Manager, it will
successfully create the connection, it will just not do anything. Since
the integration with Netplan, this scenario will start to fail and might
be surprising for users.

Also, when creating an ethernet connection through NM, the interface ID
in Netplan will end up being something like
NM-030c95e3-b025-4a1e-905f-d3d92acfd8b9. The VLAN link is supposed to
point to that interface name in the Netplan configuration. But as
translation from keyfile to YAML are handled isolated from the rest of
the netplan state, we don't know what netdef ID no point to when doing
the translation. Doing it properly would require loading and analyzing
all the Netplan state before emitting YAML for keyfiles.