abi: add the new netdef field "tunnel_private_key_flags"
The enum basically shadows NM's NMSettingSecretFlags.

It will be used to store the Wireguard's private key flag if defined.
The main reason for that is that the private key might not be present in
the keyfile. It might the owned by an external agent that handles
credentials. This is Network Manager specific. See LP: #2024661.