parse: handle redefinition of a place holder netdef
If we find an interface with the same name of a place holder created
earlier, we just change the place holder type to the type of the
new interface and use it to store the configuration.

Without this logic the parser would fail if it finds an interface with
the same name of the place holder but of a different type.