parse-nm: handle cloned-mac-address special cases
The property "cloned-mac-address" might contain one of these special
values: "preserve", "permanent", "random" and "stable". Currently we
accept only a valid MAC address as value.

This change will leave the cloned-mac-address property in the
networkmanager.passthrough section if it contains one of these values
and will try to validate it otherwise.

Related to LP: #2026230