netplan:cli:state: new module to collect network configuration
state.py is responsible for collecting and storing the network
configuration. It will also be responsible for detecting differences
between the system's and netplan configuration in the future.

The main motivation for this new submodule is the centralization of
state management.
The code used by "netplan status" to read and aggregate the system's
network configuration and the code used by "netplan get" to read the
Netplan's configuration was moved to this module.