tests:integration: Adopt for systemd v254
systemd-networkd v245 seems to be more picky about moving an interface from "configuring" to "configured" state.
It waits for an IPv6 RA to be received. So we need to define "accept-ra: false" in Netplan tests to ignore that,
or make our test environment (dnsmasq) to actually provide RAs. I went with the latter.

We still see weird error messages from 'networkctl', which seems new, but they do not fail tests. E.g.:
Failed to query link bit rates: Invalid argument
Failed to query link DHCP leases: Invalid argument

I have the impression that those error messages are thrown while the interface is still in "configuring" state,
i.e. while waiting for the RA to be received/processed. We're calling 'networkctl status IFACE' in a loop and
the error message vanishes when calling the command after a few seconds.