ctests: add a new cmocka test file for the NM backend
The only test at the moment is related to a memory leak plugged in the
function netplan_state_finish_nm_write() that happens when the state is
empty.

A couple of non-related functions were renamed so they will not collide
with functions with the same name defined in another file.