ctests: add a new cmocka openvswitch test file
The only test at the moment check that write_ovs_bond_interfaces()
returns NULL when netdef->bridge is NULL. It's also used to check if the
function doesn't leak memory in this case.

The function write_bridge_params() was renamed so the name will not
conflict with another function defined in another file.