commands/status: query routes from all the tables
We were listing routes from only the main table. The new diff
functionality will need all routes in order to calculate the diff
properly.

The extra routes are not displayed by default.

A new parameter, --verbose, was added and the extra routes will be
displayed when netplan status is called with it.