netplan: add support for WPA3-Enterprise
PMF (Protected Management Frames) is required by WPA3 and was
already implicitly set to "required" for WPA3-Personal (via SAE).

Network Manager will enable different EAP methods simultaneously when we
set it to "eap", such as WPA-EAP and WPA-EAP-SHA256. NM doesn't allow
the user to set the method as "eap-sha256" only.

These changes add two new EAP methods to Netplan: "eap-sha256" and
"eap-suite-b-192". They are both used with WPA3-Enterprise.

PMF is mandatory when using "eap-suite-b-192" so it's implicitly set to
"required". It's implicitly set to "optional" when eap-sha256 is used.