state_diff: add the new state_diff submodule
The state_diff submodule is responsible for comparing system's and
netplan's states and calculating the differences.

The new class NetplanDiffState must be instantiated with both states.
The method get_full_state() will return both system's and netplan's
states without calculating the differences.

The diff calculation will be added in the next commits.