parse: improve the parsing of access-points (LP: #1809994)
When the parser requires a second pass, it will report access-points it
already parsed as duplicates and error out.

With this change, access-points will be stored in a local hash table,
which will be used to find duplicates in the same netdef, and moved to
the netdef->access_points hash table in the end. Access-points already
present in the netdef hash table will be considered already parsed and
skipped.

Add a test to check the problem is fixed and another one to test if
merging the access-points when the same interface is present in multiple
files is working properly.