wireguard: ignore empty endpoints
The Network Manager GUI will emit 'Endpoint=' if it's not specified when
creating the tunnel. libnetplan is generating 'endpoint: ""' in this case
and the parser is returning an error.

In nm-parse, if Endpoint is either empty or NULL we'll just not read it. In
the parser, if we find 'endpoint: ""', we'll just ignore it as if it
were not defined.

See LP: #2038811