wifi: replace the previously defined AP with the new one
If we find an AP with the same name in the same interface, we drop the
first one and use the new one. This is done to maintain the documented
behavior of respecting the order in which the files are parsed. We still
need to implement support for merging AP settings.