tests/ctests: Build ctests separately
At present the ctests are not properly linked with the built code,
this causes issues when there are interdependencies between the
modules.

The meson build system does not appear to have support for
internal build order or dependncies, so the only way I have found
to make this work is to break out the build of the ctests as a
separate project.

This is required to avoid undefined symbols in the test code for
the next patch in this series.

Signed-off-by: Frode Nordahl <frode.nordahl@canonical.com>