ovs: Support consuming OVS from snap
At present netplan hard codes the location of the `ovs-vsctl`
binary.  This does not work well when trying to integrate with
snaps.

Determine path of `ovs-vsctl` at runtime by checking for both
'/snap/bin/ovs-vsctl' and '/usr/bin/ovs-vsctl' with a preference
of the former.

This allows making use of the Netplan Open vSwitch integration
with for example the MicroOVN snap (and its consumers such as
MicroCloud).

TODO: Figure out how to hanlde the ovsdb-server.service unit check
TODO: Add test cases that checks the snap workflow

Signed-off-by: Frode Nordahl <frode.nordahl@canonical.com>