state_diff: fix filtering of host scoped routes
When an interface has multiple IPs in the same subnet, there will be
host scoped routes for each of them and the "src" address will be one of
the IPs for all of them.

The filtering of this type of route was not working properly and
including the extra routes in the diff.

This change generalizes the verification to see if the destination
address is one of the system own addresses instead of comparing it with
the src address.

While here, fix the data structure passed to the filtering method. It
should be a list but a dictionary was being passed to it.