ctests: stop including C files in the test files
With these changes, a copy of libnetplan without static functions will be
build. All the test files will be linked against this separate
library. With that, there is no need to include the C files in the test
files anymore.

To make this possible, all the static keywords in function declarations
were replaced with a macro that will be defined at compilation time.