state/status: add support for VRF members and interface kind
"netplan status" will now also show information about what interfaces are
members of a VRF and for each interface show what is the VRF it's
associated to.

It will also look for the Kind property to be more specific about the
interface type. For example, instead of showing a dummy interface as
being an ethernet device it will show dummy.