sriov: move the udev logic to a service unit
netplan apply will be called multiple times in parallel when the udev
rules matches multiple device. That might cause problems during boot.

Move the call to "netplan apply --sriov-only" to a service so it will be
called just once for all the interfaces. The sriov-rebind service must
wait for this new service to finish before start running.

This change was tests with real hardware and appears to work fine.