sriov: check the eswitch mode before trying to change it
The eswitch mode will be modified with devlink regardless its current
mode. It causes issues when VF LAG is enabled in the PF. This particular
problem will only happen on Mellanox interfaces (hopefully) due to the
VF LAG feature.

Retrieving the eswitch mode with devlink seems to not work on all NICs.
If it fails to get the information, the new method will return
'undetermined' and we set the eswitch mode anyway.

In a setup with Mellanox NICs with eswitch mode set to switchdev and VF
LAG enabled, netplan apply will break the setup even if there is nothing
to change in the SR-IOV NICs. This change will prevent it to happen.