sriov_rebind: cooperate with VF LAG activation
When a PF is part of a bond interface, the mlx5 driver will try to
activate the VF LAG feature. While it happens, VFs can't be bound or the
process will fail.

This change adds a verification step to netplan rebind so, if the PF is
part of a bond, it will wait for a few seconds until the VF LAG feature
is reported as active by the driver.

This is done through the debugfs interface provided by the mlx5 driver
in newer version of the kernel (5.19+).