state/status: Rename Members to Interfaces
The name Interfaces is used in the Netplan configuration already.

Also, check if the interface 'Type' is being reported as 'none'. I have
an active OpenVPN tunnel and networkctl return 'none' as type, but the
'Kind' property is 'tun'. So use the Kind property in these cases. Also
add a test case for that.