sriov_rebind: netplan rebind --debug setup
All the subcommands support the parameter --debug but it's not really
implemented (only the global netplan --debug is).

The main reason for that is because the global --debug also enables glib
debugging messages and I'd like to not have all that noise in the output
of netplan rebind as --debug will be used by default in the systemd
service unit.

But the real reason for that is because it's really tricky to get the
messages from the root logger from stderr in unit tests. Changing the
root logger to log them to stdout will break many tests.