parser: accept special options for MAC address
Besides the actual MAC address value, Network Manager and networkd
also accept special options to control the assignment of a MAC address.

The options supported are: permanent, random, stable and preserve.

All these options are supported by Network Manager. Only 'permanent'
(which will be translated to 'persistent') and 'random' are supported by
networkd and will be added to the .link file.