sriov_apply: execute apply --sriov-only before network-pre.target
When activating VF LAG and at the same time changing the e-switch mode
to switchdev, the e-switch change MUST be performed before the bonding.
Trying to change it after the bond is created will fail and the driver
will report that the e-switch is busy.

In fact, all the SR-IOV initial setup must happen before the networking
configuration.

Change the order in which we call netplan apply --sriov-only to happen
before network-pre.target.

Also, add DefaultDependencies=no to the service to prevent the creation
of dependency cycles.